+++ mac_os_x/Interface_Source/CLIPSTextDocument.m
@@ -116,7 +116,7 @@
 - (void) textViewDidChangeSelection: (NSNotification *) aNotification
   {
    NSRange selectionRange;
-   unsigned int cursorLocation;
+   NSUInteger cursorLocation;
    unichar characterToCheck;
    unsigned short nestingDepth;
    NSString *theText = [textView string];
@@ -505,10 +505,10 @@
 - (IBAction) balance: (id) sender
   {
    NSRange selectionRange;
-   unsigned int leftMiddle, rightMiddle, textLength;
+   NSUInteger leftMiddle, rightMiddle, textLength;
    unichar characterToCheck;
    int count, leftCount, rightCount;
-   unsigned int i;
+   NSUInteger i;
    BOOL endReached;
    NSString *theText = [textView string];
    
@@ -655,11 +655,11 @@
 /* and right parentheses is balanced.                   */
 /********************************************************/
 - (void) balanceIt: (NSString *) theText
-         leftMiddle: (unsigned int) leftMiddle
-         rightMiddle: (unsigned int) rightMiddle
+         leftMiddle: (NSUInteger) leftMiddle
+         rightMiddle: (NSUInteger) rightMiddle
          leftCount: (int) leftCount
          rightCount: (int) rightCount
-         textLength: (unsigned int) textLength
+         textLength: (NSUInteger) textLength
   {
    unichar characterToCheck;
    BOOL beginningReached = NO;
@@ -730,8 +730,8 @@
   {
    NSRange selectionRange, rangeOfLine, tempRange;
    unsigned int numberOfLines = 0, i;
-   unsigned int tempLocation, maxSelectedRange;
-   unsigned locationOfFirstLine, locationOfLastLine;
+   NSUInteger tempLocation, maxSelectedRange;
+   NSUInteger locationOfFirstLine, locationOfLastLine;
    NSString *theText = [textView string];
       
    /*==============================================*/
@@ -837,8 +837,8 @@
    NSRange selectionRange, rangeOfLine, tempRange;
    unsigned int numberOfLines = 0, i;
    unsigned int numberOfCharactersDeleted = 0;
-   unsigned int tempLocation, maxSelectedRange;
-   unsigned locationOfFirstLine, locationOfLastLine;
+   NSUInteger tempLocation, maxSelectedRange;
+   NSUInteger locationOfFirstLine, locationOfLastLine;
    NSString *theText = [textView string];
       
    /*==============================================*/
